Fix bad map cache dir generation
Currently it uses map.name to generate the map cache dir for saving the permissibility region. Unfortunately, map.name returns "map", rather than the image file of the map. It doesn't look like there's a simple way to get the image file path from the map topic, so I suggest using the map resolution as a unique identifier (maybe chuck in the scale as well?) it's certainly not foolproof, but it's a start.